Refactoring: `metadata.py`
This commit add a refactoring for `metadata.py`.

This python package name conflicts with the new `/api/v1/metadata`
naming.

The current implementation on `repository_service_tuf_api/metadata.py`
package are core initialization required across other implementations
and were moved to `repository_service_tuf_api/__init__.py`.

It doesn't affect or change the user usage of the RSTUF API.
